Revert "Merge feat/managers into fix/v6-consolidation"
This reverts commit 99845f263bddcdd41edf2d9fb8f602e379b2c17d.
