+++ frontend/src/components/Table/HolderTable.js
@@ -12,25 +12,12 @@ import { compareByProperty } from "@utils";
 
 import { useManageHolders } from "@hooks";
 
+import { HOLDER_HEAD_ROWS } from "@static";
+
 import "@style/Table/HolderTable.css";
 
 const HolderTable = ({ badge, isManager }) => {
-    const [headRows, setHeadRows] = useState({
-        ethereum_address: {
-            label: "Holder",
-            sortable: true,
-            method: "",
-            align: "left",
-            width: "80%"
-        },
-        balance: {
-            label: "Balance",
-            sortable: true,
-            method: "",
-            align: "right",
-            width: "20%"
-        }
-    });
+    const [headRows, setHeadRows] = useState(HOLDER_HEAD_ROWS);
     
     // The new holder objects being created by the user.
     const [newHolders, setNewHolders] = useState([]);
@@ -63,14 +50,14 @@ const HolderTable = ({ badge, isManager }) => {
     }
 
     const onAddNew = () => {
-        setNewHolders(newHolders => ([{ethereum_address: "", pendingAmount: ""}, ...newHolders]))
+        setNewHolders([{ethereum_address: "", pendingAmount: ""}, ...newHolders])
     }
 
     // If they're a new holder, update the newHolders object,
     // If they're a current holder, we need to update the balanceChanges object to give them a new balance.
     const onAmountChange = (e, index, isActive) => {
         isActive ?
-            setBalanceChanges(balanceChanges => ({...balanceChanges, [index]: {...badge.users[index], pendingAmount: e.target.value }})) :
+            setBalanceChanges({...balanceChanges, [index]: {...badge.users[index], pendingAmount: e.target.value }}) :
             setNewHolders(newHolders => newHolders.map((holder, i) => i === index ? {...holder, pendingAmount: e.target.value} : holder))
     }
     
@@ -82,7 +69,7 @@ const HolderTable = ({ badge, isManager }) => {
     // If they're a current holder, we need to update the balanceChanges object to give them a new balance of 0.
     const onDelete = (index, isActive) => {
         isActive ?
-            setBalanceChanges(balanceChanges => ({...balanceChanges, [index]: {...badge.users[index], pendingAmount: "0" }})) :
+            setBalanceChanges({...balanceChanges, [index]: {...badge.users[index], pendingAmount: "0" }}) :
             setNewHolders(newHolders => newHolders.filter((holder, i) => i !== index))
     }
 
@@ -122,10 +109,11 @@ const HolderTable = ({ badge, isManager }) => {
                 body="When you add a Holder, they will appear on the list here."
             />}
 
-            {badge && (badge.users.length !== 0 || newHolders.length !== 0) && <TableContainer className="table">
-                <Table>
+            {badge && (badge.users.length !== 0 || newHolders.length !== 0) && <div id="holder__table">
+                <TableContainer>
+                  <Table>
                     <TableHead>
-                        <TableRow>
+                      <TableRow>
                         {Object.keys(headRows).map((key) => (
                             <TableSortHead
                                 key={key}
@@ -137,7 +125,7 @@ const HolderTable = ({ badge, isManager }) => {
                                 width={headRows[key].width}
                             />
                         ))}
-                        </TableRow>
+                      </TableRow>
                     </TableHead>
 
                     <TableBody>
@@ -186,8 +174,9 @@ const HolderTable = ({ badge, isManager }) => {
                         ))}
 
                     </TableBody>
-                </Table>
-            </TableContainer>}
+                  </Table>
+                </TableContainer>
+            </div>}
         </>
     )
 }
