+++ frontend/src/components/Table/index.js
@@ -1,5 +1,4 @@
 export { BadgeTable } from './BadgeTable';
 export { HolderTable } from './HolderTable';
-export { ManagerTable } from './ManagerTable';
 export { ObjectTable } from './ObjectTable';
 export { TableSortHead } from './TableSortHead';
