+++ frontend/src/static/constants/index.js
@@ -1,3 +1,3 @@
-export { pfpEmojis, getRandomEmoji } from './constants'
+export { BADGE_HEAD_ROWS, HOLDER_HEAD_ROWS, pfpEmojis, getRandomEmoji, badgeDrawerSelectActions } from './constants'
 export { ERRORS } from './errors'
 export { HOME_LINKS, IPFS_GATEWAY_URL } from './links'
